# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platform Identity \u2014 X.TLD application integration contract",
    "version": "2.0.0",
    "description": "Identity is the sole platform authority for users, provider identities, tenants, memberships, and SSO/application browser sessions. Trusted server admission uses trustedCIDR (default) or configured secret/dual compatibility mode. Directory operations also require an application-session bearer actor. SUPER_ADMIN uses provider-proven session authority unless explicitly granted or removed by another Super Admin; directory email never grants global access. See PLATFORM_CONTRACT.md for rollout, role, online introspection, and data preservation rules."
  },
  "servers": [
    {
      "url": "https://identity.X.TLD",
      "description": "X.TLD is the configured PARENT_DOMAIN (placeholder, normative)."
    }
  ],
  "paths": {
...
    "/api/directory/users": {
      "post": {
        "operationId": "ensureDirectoryUser",
        "summary": "Idempotently ensure a central user by email (CIDR-trusted server only)",
        "description": "Requires trusted server admission and a current SUPER_ADMIN app-session bearer. Creates the user if no user holds the normalised email, otherwise returns the existing one; concurrency-safe, and a repeated `idempotencyKey` returns the UID the first call produced. A pre-created user is `claimed: false` until a trusted-provider login with a matching verified email attaches an identity \u2014 untrusted providers can never claim a UID by asserted email. No client-secret header or body field is accepted on any directory endpoint, in any deployment mode.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectoryEnsureRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The ensured user (minimal fields only \u2014 never identities, sessions, codes, or OAuth credentials).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryUser"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Caller's resolved IPv4 peer is not allowlisted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ]
      },
      "get": {
        "operationId": "searchDirectoryUsers",
        "summary": "Search central users (CIDR-trusted server only)",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "description": "Literal substring matched against email and displayName.",
            "schema": {
              "type": "string",
              "maxLength": 255
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Keyset cursor: the `nextCursor` of the previous page; 0 for the first page.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of users.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryPage"
                }
              }
            }
          },
          "400": {
            "description": "limit or cursor out of bounds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Caller's resolved IPv4 peer is not allowlisted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ],
        "description": "Requires trusted server admission and a current SUPER_ADMIN app-session bearer. "
      }
    },
    "/api/directory/users/{iUserId}": {
      "get": {
        "operationId": "getDirectoryUser",
        "summary": "Fetch one central user by iUserId (CIDR-trusted server only)",
        "parameters": [
          {
            "name": "iUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryUser"
                }
              }
            }
          },
          "404": {
            "description": "No such user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Caller's resolved IPv4 peer is not allowlisted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ],
        "description": "Requires trusted server admission and a current SUPER_ADMIN app-session bearer. "
      },
      "patch": {
        "summary": "Update user display name (SUPER_ADMIN)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryUser"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or revoked actor session"
          },
          "403": {
            "description": "Untrusted server or unauthorized actor"
          },
          "409": {
            "description": "Conflicting record or last administrator removal"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 255
                  }
                },
                "required": [
                  "displayName"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ],
        "parameters": [
          {
            "name": "iUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ]
      }
    },
    "/api/sessions/introspect": {
      "post": {
        "summary": "Read current session and tenant authority (trusted server)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Introspection"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or revoked actor session"
          },
          "403": {
            "description": "Untrusted server or unauthorized actor"
          },
          "409": {
            "description": "Conflicting record or last administrator removal"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSession"
              }
            }
          }
        }
      }
    },
    "/api/sessions/revoke": {
      "post": {
        "summary": "Revoke central application session (trusted server)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revoked": {
                      "const": true
                    }
                  },
                  "required": [
                    "revoked"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or revoked actor session"
          },
          "403": {
            "description": "Untrusted server or unauthorized actor"
          },
          "409": {
            "description": "Conflicting record or last administrator removal"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSession"
              }
            }
          }
        }
      }
    },
    "/api/sessions/select-tenant": {
      "post": {
        "summary": "Persist a permitted enabled tenant in the central session",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "selectedTenantId": {
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "selectedTenantId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or revoked actor session"
          },
          "403": {
            "description": "Untrusted server or unauthorized actor"
          },
          "409": {
            "description": "Conflicting record or last administrator removal"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "pattern": "^[0-9a-f]{64}$"
                  },
                  "iTenantId": {
                    "anyOf": [
                      {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 9007199254740991
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "token",
                  "iTenantId"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/runtime/tenants/{iTenantId}": {
      "get": {
        "summary": "Read tenant enablement (trusted service only)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "iTenantId": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 9007199254740991
                    },
                    "bEnabled": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "iTenantId",
                    "bEnabled"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or revoked actor session"
          },
          "403": {
            "description": "Untrusted server or unauthorized actor"
          },
          "409": {
            "description": "Conflicting record or last administrator removal"
          }
        },
        "parameters": [
          {
            "name": "iTenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ]
      }
    },
    "/api/directory/tenants": {
      "get": {
        "summary": "List current actor tenants",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tenants": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Tenant"
                      }
                    }
                  },
                  "required": [
                    "tenants"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or revoked actor session"
          },
          "403": {
            "description": "Untrusted server or unauthorized actor"
          },
          "409": {
            "description": "Conflicting record or last administrator removal"
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ]
      },
      "post": {
        "summary": "Create tenant (SUPER_ADMIN)",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or revoked actor session"
          },
          "403": {
            "description": "Untrusted server or unauthorized actor"
          },
          "409": {
            "description": "Conflicting record or last administrator removal"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "slug": {
                    "type": "string",
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "maxLength": 100
                  }
                },
                "required": [
                  "name",
                  "slug"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ]
      }
    },
    "/api/directory/tenants/{iTenantId}": {
      "patch": {
        "summary": "Update own tenant; enablement requires SUPER_ADMIN",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or revoked actor session"
          },
          "403": {
            "description": "Untrusted server or unauthorized actor"
          },
          "409": {
            "description": "Conflicting record or last administrator removal"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "slug": {
                    "type": "string",
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "maxLength": 100
                  },
                  "bEnabled": {
                    "type": "boolean"
                  }
                },
                "required": [],
                "additionalProperties": false,
                "minProperties": 1
              }
            }
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ],
        "parameters": [
          {
            "name": "iTenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ]
      }
    },
    "/api/directory/tenants/{iTenantId}/memberships": {
      "get": {
        "summary": "List own tenant memberships (tenant admin)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memberships": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Membership"
                      }
                    }
                  },
                  "required": [
                    "memberships"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or revoked actor session"
          },
          "403": {
            "description": "Untrusted server or unauthorized actor"
          },
          "409": {
            "description": "Conflicting record or last administrator removal"
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ],
        "parameters": [
          {
            "name": "iTenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ]
      }
    },
    "/api/directory/tenants/{iTenantId}/memberships/{iUserId}": {
      "put": {
        "summary": "Edit member profile and role; global grants require Super Admin",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or revoked actor session"
          },
          "403": {
            "description": "Untrusted server or unauthorized actor"
          },
          "409": {
            "description": "Conflicting record or last administrator removal"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "enum": [
                      "SUPER_ADMIN",
                      "TENANT_ADMIN",
                      "USER"
                    ]
                  },
                  "bEnabled": {
                    "type": "boolean"
                  },
                  "displayName": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 255
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255
                  }
                },
                "required": [
                  "role",
                  "bEnabled"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ],
        "parameters": [
          {
            "name": "iTenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "iUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DirectoryEnsureRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "additionalProperties": true,
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255
          },
          "displayName": {
            "type": "string",
            "maxLength": 255
          },
          "idempotencyKey": {
            "type": "string",
            "maxLength": 128
          }
        }
      },
      "DirectoryUser": {
        "type": "object",
        "required": [
          "iUserId",
          "email",
          "displayName",
          "claimed"
        ],
        "additionalProperties": false,
        "properties": {
          "iUserId": {
            "type": "integer"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "claimed": {
            "type": "boolean",
            "description": "True once any login method is attached \u2014 a trusted-provider login with a matching verified email claims a pre-created UID; untrusted providers never can."
          }
        },
        "description": "Deliberately minimal: never identities, sessions, codes, or OAuth credentials."
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Generic; 403 bodies never disclose the policy or the resolved peer."
          },
          "correlationId": {
            "type": "string",
            "description": "Present on 403s; matches the server log line carrying the resolved peer IP."
          }
        }
      },
...
    },
    "securitySchemes": {
      "AppSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque central application session plus trusted server admission"
      }
    }
  }
}